Revert "Graphics (#8)"
This reverts commit 5085e99fa51a901a91d96a52ecc1f9d3d1adee9b.
