+++ ToDoList/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\User\\AppData\\Roaming\\NetBeans\\13\\build.properties
